+++ src/index.ts
@@ -194,6 +194,11 @@ class TheAlgorithm {
         ]
 
         await Promise.all([...initialLoads, ...secondaryLoads]);
+
+        // Now that all data has arrived, go back over and do the slow stuff.
+        await Toot.setDependentProps(this.feed, TRIGGER_FEED + " DEEP", true);
+        updatePropertyFilterOptions(this.filters, this.feed, await MastoApi.instance.getUserData());
+        await this.scoreAndFilterFeed();
         this.finishFeedUpdate();
     }