+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Defer processing of the most expensive Toot properties (trendingLinks, mostly, but other tag stuff too) until all data has arrived
 * `fetchHomeFeed()` now infers the `maxTootedAt` param from the cached `HOME_TIMELINE` toots
 * New private property `TheAlgorithm.homeFeed: Toot[]`
 * Rename `Config.backgroundLoadIntervalSeconds`